Working copy ready for review
This version of the code should successfully calculate metrics
which will be more fully described when the code is verified with
more rigerous analysis

However briefly,
This script trains a model with an initial sample size of 10 samples
per subject.  It then iteratively: tests the next 10 samples against
the model, takes all successfully classified samples and add them
to the model, retrain the model, repeat with next 10 samples.

The above procedure is repeated for k in {1,3,5,7,9} and for
initial training sizes in multiples of 10 from 10 to 390